add results actuator for spatial search
